+++ Sources/Leaf/Service/Provider.swift
@@ -45,42 +45,3 @@ public final class LeafProvider: Provider {
     /// See Service.Provider.boot
     public func boot(_ container: Container) throws { }
 }
-
-
-// MARK: View
-
-public struct View: Codable {
-    /// The view's data.
-    public let data: Data
-
-    /// Create a new View
-    public init(data: Data) {
-        self.data = data
-    }
-
-    /// See Encodable.encode
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.singleValueContainer()
-        try container.encode(data)
-    }
-
-    /// See Decodable.decode
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        try self.init(data: container.decode(Data.self))
-    }
-}
-
-
-public protocol ViewRenderer {
-    /// Renders a view using the supplied encodable context and worker.
-    func make<E>(_ path: String, _ context: E) throws -> Future<View>
-        where E: Encodable
-}
-
-extension ViewRenderer {
-    /// Create a view with null context.
-    public func make(_ path: String) throws -> Future<View> {
-        return try make(path, nil as String?)
-    }
-}
